Split metadata writer: rename to metadata.rs, extract heif.rs
The writer file now covers EXIF + IPTC + XMP across six container formats,
so `exif.rs` no longer reflects its scope. Renamed to `metadata.rs`.

The HEIF/ISO-BMFF atom surgery (~330 lines of mp4-atom work that XMP
Toolkit can't handle) lived in the same file but is its own concern;
moved into a dedicated `heif.rs` module. `metadata.rs` retains the
shared `apply_to_xmp` helper and dispatches to `heif::insert_xmp` for
HEIC / HEIF / AVIF paths.

No behavior change.
