+++ src/download/metadata.rs
@@ -1,20 +1,22 @@
-//! Embedded metadata (XMP + native EXIF/IPTC reconciliation) via Adobe's
-//! XMP Toolkit.
+//! Embedded metadata (XMP + native EXIF/IPTC reconciliation).
 //!
-//! The writer runs through [`xmp_toolkit::XmpFile`], which vendors Adobe's
-//! reference XMPFiles implementation. One code path covers JPEG, HEIC, PNG,
-//! TIFF, MP4, MOV, and more — whatever kei downloads from iCloud ends up with
-//! the same metadata embedded in its file bytes.
+//! For JPEG / PNG / TIFF / MP4 / MOV, the writer runs through
+//! [`xmp_toolkit::XmpFile`] — Adobe's vendored XMPFiles implementation, which
+//! also reconciles XMP with native EXIF/IPTC blocks so consumers that read
+//! only EXIF still see values like `Rating`, GPS, and `DateTimeOriginal`.
 //!
-//! XMP Toolkit also reconciles XMP with native EXIF/IPTC blocks on formats
-//! that carry them (notably JPEG), so a consumer reading only EXIF still
-//! sees values like `Rating`, GPS, and `DateTimeOriginal`.
+//! HEIC / HEIF / AVIF have no XMP Toolkit handler, so those formats route
+//! through [`super::heif`], which edits the ISO-BMFF container directly with
+//! [`mp4_atom`]. Both paths build the XMP packet via the same
+//! [`apply_to_xmp`] helper, so the embedded content is identical.
 
 use std::path::Path;
 use std::sync::Once;
 
 use anyhow::{Context, Result};
 use xmp_toolkit::{xmp_ns, OpenFileOptions, XmpFile, XmpMeta, XmpValue};
+
+use super::heif;
 
 /// Custom XMP namespace for kei-specific fields that don't fit standard
 /// schemas (`hidden`, `archived`, `mediaSubtype`, `burstId`). Consumers that
@@ -112,10 +114,6 @@ impl MetadataWrite {
 /// Write the requested metadata into the file's XMP packet, with EXIF/IPTC
 /// reconciliation where the container supports it.
 ///
-/// HEIC/HEIF routes through `libheif` because Adobe's vendored XMPFiles has
-/// no HEIF handler. Everything else (JPEG, PNG, TIFF, MP4, MOV, …) uses
-/// `xmp_toolkit`.
-///
 /// Atomic: we copy the input to a sibling `.meta-tmp`, patch it in place,
 /// then rename over the target. A crash mid-write leaves the original
 /// untouched.
@@ -123,23 +121,11 @@ pub(crate) fn apply_metadata(path: &Path, write: &MetadataWrite) -> Result<()> {
     if write.is_empty() {
         return Ok(());
     }
-    if is_heif_path(path) {
+    if heif::is_heif_path(path) {
         apply_metadata_heif(path, write)
     } else {
         apply_metadata_xmp_toolkit(path, write)
     }
-}
-
-/// Whether this path's extension is HEIF / HEIC / HIF / AVIF — formats
-/// that XMP Toolkit's bundled handlers can't open but libheif can.
-fn is_heif_path(path: &Path) -> bool {
-    path.extension()
-        .and_then(|e| e.to_str())
-        .map(|e| {
-            let lower = e.to_ascii_lowercase();
-            matches!(lower.as_str(), "heic" | "heif" | "hif" | "avif")
-        })
-        .unwrap_or(false)
 }
 
 fn apply_metadata_xmp_toolkit(path: &Path, write: &MetadataWrite) -> Result<()> {
@@ -301,7 +287,7 @@ fn apply_metadata_heif(path: &Path, write: &MetadataWrite) -> Result<()> {
     // Preserve any XMP the file already carries (e.g. Apple Live Photo or
     // depth markers) by parsing it into the XmpMeta we mutate. If parsing
     // fails or there's no existing XMP, start from an empty packet.
-    let existing_xmp_bytes = extract_xmp_bytes_from_heif(&input);
+    let existing_xmp_bytes = heif::extract_xmp_bytes(&input);
     let mut meta = existing_xmp_bytes
         .as_deref()
         .and_then(|bytes| std::str::from_utf8(bytes).ok())
@@ -310,7 +296,7 @@ fn apply_metadata_heif(path: &Path, write: &MetadataWrite) -> Result<()> {
     apply_to_xmp(&mut meta, write)?;
     let xmp_bytes = meta.to_string().into_bytes();
 
-    let new_bytes = insert_xmp_into_heif(&input, &xmp_bytes)
+    let new_bytes = heif::insert_xmp(&input, &xmp_bytes)
         .with_context(|| format!("Inserting XMP into HEIC {}", path.display()))?;
 
     let mut tmp_name = path.file_name().unwrap_or_default().to_os_string();
@@ -322,385 +308,6 @@ fn apply_metadata_heif(path: &Path, write: &MetadataWrite) -> Result<()> {
         .with_context(|| format!("Renaming {} -> {}", tmp_path.display(), path.display()))?;
     tracing::debug!(path = %path.display(), "Applied HEIC metadata");
     Ok(())
-}
-
-/// Locate the XMP packet bytes embedded in a HEIC file, if any. Returns the
-/// raw RDF/XML payload referenced by the first `mime`-type item with
-/// content_type `"application/rdf+xml"`. Used by the write path to preserve
-/// existing XMP on rewrite (symmetric with xmp_toolkit's `file.xmp()`).
-fn extract_xmp_bytes_from_heif(bytes: &[u8]) -> Option<Vec<u8>> {
-    use mp4_atom::{Any, DecodeMaybe, FourCC, Iinf, Iloc};
-    let mut cursor: &[u8] = bytes;
-    while let Ok(Some(atom)) = Any::decode_maybe(&mut cursor) {
-        let Any::Meta(meta) = atom else { continue };
-        let iinf = meta.get::<Iinf>()?;
-        let iloc = meta.get::<Iloc>()?;
-        let xmp_entry = iinf.item_infos.iter().find(|e| {
-            e.item_type == Some(FourCC::new(b"mime"))
-                && e.content_type.as_deref() == Some("application/rdf+xml")
-        })?;
-        let loc = iloc
-            .item_locations
-            .iter()
-            .find(|l| l.item_id == xmp_entry.item_id)?;
-        if loc.construction_method != 0 {
-            return None;
-        }
-        let extent = loc.extents.first()?;
-        let start = loc.base_offset.saturating_add(extent.offset) as usize;
-        let end = start + extent.length as usize;
-        if end > bytes.len() {
-            return None;
-        }
-        return Some(bytes[start..end].to_vec());
-    }
-    None
-}
-
-/// Surgically insert (or replace) the XMP `mime` item inside a HEIC file.
-///
-/// The HEIC container is ISO-BMFF — a sequence of top-level atoms. XMP lives
-/// inside the `meta` atom as an item with `item_type = "mime"` and
-/// `content_type = "application/rdf+xml"`. We store the XMP bytes in an
-/// `idat` atom inside `meta` (construction_method 1, offsets relative to
-/// idat), so the encoded image bytes in `mdat` — referenced by file-absolute
-/// offsets from other `iloc` entries — just shift by however much `meta` grew.
-fn insert_xmp_into_heif(input: &[u8], xmp: &[u8]) -> Result<Vec<u8>> {
-    use mp4_atom::{
-        Any, DecodeMaybe, Encode, FourCC, Iloc, ItemInfoEntry, ItemLocation, ItemLocationExtent,
-        Mdat,
-    };
-
-    // Record each top-level atom along with its original byte offset in the
-    // input so we can rewrite file-absolute iloc entries correctly — the
-    // existing iloc offsets point into the original mdat, and those offsets
-    // must be updated so that after re-serialization they still land on the
-    // same image bytes even though the meta box grew.
-    let mut cursor: &[u8] = input;
-    let mut atoms: Vec<Any> = Vec::new();
-    let mut original_offsets: Vec<u64> = Vec::new();
-    while !cursor.is_empty() {
-        let offset_before_atom = (input.len() - cursor.len()) as u64;
-        match Any::decode_maybe(&mut cursor)? {
-            Some(a) => {
-                atoms.push(a);
-                original_offsets.push(offset_before_atom);
-            }
-            None => anyhow::bail!(
-                "unparseable tail at offset {} of {}",
-                offset_before_atom,
-                input.len()
-            ),
-        }
-    }
-
-    let meta_idx = atoms
-        .iter()
-        .position(|a| matches!(a, Any::Meta(_)))
-        .ok_or_else(|| anyhow::anyhow!("HEIC has no meta box"))?;
-
-    // Step 1: locate and drop the trailing mdat that a prior kei write
-    // appended (if any) so we don't accumulate stale XMP payloads on
-    // re-sync. We identify it by: (a) the existing XMP iloc entry's
-    // extent range, (b) it sitting past the image-data mdat, (c) no
-    // other iloc entry pointing into it.
-    let stale_mdat_idx = locate_stale_kei_mdat(&atoms, &original_offsets, meta_idx);
-
-    // Step 2: remove the XMP entries from iinf and iloc. No idat mutation
-    // needed because we don't put XMP in idat — it goes into a dedicated
-    // mdat at end of file (see step 3 below).
-    if let Any::Meta(meta) = &mut atoms[meta_idx] {
-        let removed_ids = remove_existing_xmp_items(meta);
-        if let Some(iloc) = meta.get_mut::<Iloc>() {
-            iloc.item_locations
-                .retain(|loc| !removed_ids.contains(&loc.item_id));
-        }
-    }
-
-    // Step 3: drop the stale mdat atom (indexes shift, recompute meta_idx
-    // relative to the surviving atoms).
-    let meta_idx = if let Some(stale) = stale_mdat_idx {
-        atoms.remove(stale);
-        original_offsets.remove(stale);
-        if stale < meta_idx {
-            meta_idx - 1
-        } else {
-            meta_idx
-        }
-    } else {
-        meta_idx
-    };
-
-    // Step 4: reserve the iinf + iloc entries for the new XMP. The iloc
-    // offset is the file offset our appended mdat's DATA will have in the
-    // re-serialized output. mp4-atom encodes Iloc at fixed width regardless
-    // of offset value, so we can append the mdat atom first, compute the
-    // resulting running offsets, then populate the iloc offset.
-    let new_item_id = {
-        let Any::Meta(meta) = &atoms[meta_idx] else {
-            unreachable!()
-        };
-        next_free_item_id(meta)
-    };
-
-    atoms.push(Any::Mdat(Mdat { data: xmp.to_vec() }));
-    let new_mdat_idx = atoms.len() - 1;
-
-    // Insert placeholder iloc entry (offset=0) and iinf entry so that running
-    // offsets reflect the final meta size.
-    {
-        let Any::Meta(meta) = &mut atoms[meta_idx] else {
-            unreachable!()
-        };
-        push_iinf_entry(
-            meta,
-            ItemInfoEntry {
-                item_id: new_item_id,
-                item_protection_index: 0,
-                item_type: Some(FourCC::new(b"mime")),
-                item_name: String::new(),
-                content_type: Some("application/rdf+xml".to_string()),
-                content_encoding: Some(String::new()),
-                item_not_in_presentation: false,
-            },
-        );
-        push_iloc_entry(
-            meta,
-            ItemLocation {
-                item_id: new_item_id,
-                construction_method: 0,
-                data_reference_index: 0,
-                base_offset: 0,
-                extents: vec![ItemLocationExtent {
-                    item_reference_index: 0,
-                    offset: 0,
-                    length: xmp.len() as u64,
-                }],
-            },
-        );
-    }
-
-    // Step 5: remap pre-existing file-offset iloc entries and fill in the
-    // offset for the XMP iloc entry we just pushed.
-    let new_positions = running_offsets(&atoms);
-    let xmp_file_offset = new_positions[new_mdat_idx] + header_size_of(&atoms[new_mdat_idx]);
-
-    let file_offset_map: Vec<(u64, u64, u64)> = atoms
-        .iter()
-        .enumerate()
-        .take(new_mdat_idx) // skip the mdat we just added; it has no matching original
-        .filter_map(|(idx, a)| {
-            let orig = *original_offsets.get(idx)?;
-            let size = encoded_size(a);
-            Some((orig, orig + size, new_positions[idx]))
-        })
-        .collect();
-
-    if let Any::Meta(meta) = &mut atoms[meta_idx] {
-        if let Some(iloc) = meta.get_mut::<Iloc>() {
-            remap_file_offsets(iloc, &file_offset_map);
-            // Now fill in the XMP entry's offset (last iloc entry).
-            if let Some(xmp_loc) = iloc
-                .item_locations
-                .iter_mut()
-                .find(|l| l.item_id == new_item_id)
-            {
-                if let Some(extent) = xmp_loc.extents.first_mut() {
-                    extent.offset = xmp_file_offset;
-                }
-            }
-        }
-    }
-
-    let mut out = Vec::with_capacity(input.len() + xmp.len() + 512);
-    for atom in &atoms {
-        atom.encode(&mut out)?;
-    }
-    Ok(out)
-}
-
-/// Walk existing iinf/iloc to find any previously-kei-appended XMP mdat.
-/// Criteria: an iinf entry flagged as `mime` + `application/rdf+xml`, its
-/// iloc entry references a range that lies entirely within a single trailing
-/// mdat atom, and no other iloc entry references into that atom.
-fn locate_stale_kei_mdat(
-    atoms: &[mp4_atom::Any],
-    original_offsets: &[u64],
-    meta_idx: usize,
-) -> Option<usize> {
-    use mp4_atom::{Any, FourCC, Iinf, Iloc};
-    let meta = if let Any::Meta(m) = &atoms[meta_idx] {
-        m
-    } else {
-        return None;
-    };
-    let iinf = meta.get::<Iinf>()?;
-    let iloc = meta.get::<Iloc>()?;
-
-    let xmp_item_ids: Vec<u32> = iinf
-        .item_infos
-        .iter()
-        .filter(|e| {
-            e.item_type == Some(FourCC::new(b"mime"))
-                && e.content_type.as_deref() == Some("application/rdf+xml")
-        })
-        .map(|e| e.item_id)
-        .collect();
-    if xmp_item_ids.is_empty() {
-        return None;
-    }
-
-    // For each candidate, find the atom that its extent range falls within.
-    for item_id in &xmp_item_ids {
-        let Some(loc) = iloc.item_locations.iter().find(|l| l.item_id == *item_id) else {
-            continue;
-        };
-        if loc.construction_method != 0 {
-            continue;
-        }
-        let Some(extent) = loc.extents.first() else {
-            continue;
-        };
-        let abs_start = loc.base_offset.saturating_add(extent.offset);
-        let abs_end = abs_start.saturating_add(extent.length);
-
-        // Find the atom whose original byte range fully contains the extent.
-        for (idx, atom) in atoms.iter().enumerate() {
-            if !matches!(atom, Any::Mdat(_)) {
-                continue;
-            }
-            let atom_start = original_offsets[idx];
-            let atom_end = atom_start + encoded_size(atom);
-            if abs_start < atom_start || abs_end > atom_end {
-                continue;
-            }
-            // Check that no OTHER iloc entry references this atom.
-            let other_refs = iloc.item_locations.iter().any(|other| {
-                if other.item_id == *item_id || other.construction_method != 0 {
-                    return false;
-                }
-                other.extents.iter().any(|e| {
-                    let o_start = other.base_offset.saturating_add(e.offset);
-                    o_start >= atom_start && o_start < atom_end
-                })
-            });
-            if !other_refs {
-                return Some(idx);
-            }
-        }
-    }
-    None
-}
-
-/// Byte size of an atom's box header (the length field + 4-byte kind code).
-/// mp4-atom always emits a 32-bit-length header for atoms that fit — large
-/// mdats (>4GB) would use a 16-byte header, but kei isn't going to hit that.
-fn header_size_of(_atom: &mp4_atom::Any) -> u64 {
-    8
-}
-
-/// Return a vector where entry `i` is the byte offset at which atom `i` will
-/// sit in the re-serialized output (i.e. the running sum of preceding atom
-/// sizes).
-fn running_offsets(atoms: &[mp4_atom::Any]) -> Vec<u64> {
-    let mut offsets = Vec::with_capacity(atoms.len());
-    let mut running = 0u64;
-    for atom in atoms {
-        offsets.push(running);
-        running += encoded_size(atom);
-    }
-    offsets
-}
-
-/// Translate each construction_method-0 iloc offset from "original file
-/// offset" to "new file offset", using the per-atom old_start/old_end/new_start
-/// table. An offset that falls within `[old_start, old_end)` is rebased onto
-/// `new_start` with the same intra-atom position.
-fn remap_file_offsets(iloc: &mut mp4_atom::Iloc, ranges: &[(u64, u64, u64)]) {
-    for loc in &mut iloc.item_locations {
-        if loc.construction_method != 0 {
-            continue;
-        }
-        // Some encoders put the whole file offset in `base_offset` and leave
-        // extent offsets at 0; others leave base_offset 0 and put absolute
-        // offsets on each extent. Handle both by remapping either piece that
-        // lands in a known original-atom range.
-        loc.base_offset = remap_point(loc.base_offset, ranges).unwrap_or(loc.base_offset);
-        for extent in &mut loc.extents {
-            let absolute = loc.base_offset.saturating_add(extent.offset);
-            if let Some(new_abs) = remap_point(absolute, ranges) {
-                extent.offset = new_abs.saturating_sub(loc.base_offset);
-            }
-        }
-    }
-}
-
-fn remap_point(file_offset: u64, ranges: &[(u64, u64, u64)]) -> Option<u64> {
-    for &(old_start, old_end, new_start) in ranges {
-        if file_offset >= old_start && file_offset < old_end {
-            return Some(new_start + (file_offset - old_start));
-        }
-    }
-    None
-}
-
-fn encoded_size(atom: &mp4_atom::Any) -> u64 {
-    use mp4_atom::Encode;
-    let mut sink = Vec::new();
-    atom.encode(&mut sink).ok();
-    sink.len() as u64
-}
-
-fn remove_existing_xmp_items(meta: &mut mp4_atom::Meta) -> Vec<u32> {
-    use mp4_atom::{FourCC, Iinf};
-    let mut removed = Vec::new();
-    if let Some(iinf) = meta.get_mut::<Iinf>() {
-        iinf.item_infos.retain(|entry| {
-            let is_xmp = entry.item_type == Some(FourCC::new(b"mime"))
-                && entry.content_type.as_deref() == Some("application/rdf+xml");
-            if is_xmp {
-                removed.push(entry.item_id);
-                false
-            } else {
-                true
-            }
-        });
-    }
-    removed
-}
-
-fn next_free_item_id(meta: &mp4_atom::Meta) -> u32 {
-    use mp4_atom::Iinf;
-    meta.get::<Iinf>()
-        .map(|iinf| {
-            iinf.item_infos
-                .iter()
-                .map(|e| e.item_id)
-                .max()
-                .map(|m| m + 1)
-                .unwrap_or(1)
-        })
-        .unwrap_or(1)
-}
-
-fn push_iinf_entry(meta: &mut mp4_atom::Meta, entry: mp4_atom::ItemInfoEntry) {
-    use mp4_atom::Iinf;
-    if meta.get::<Iinf>().is_none() {
-        meta.push(Iinf {
-            item_infos: Vec::new(),
-        });
-    }
-    meta.get_mut::<Iinf>().unwrap().item_infos.push(entry);
-}
-
-fn push_iloc_entry(meta: &mut mp4_atom::Meta, loc: mp4_atom::ItemLocation) {
-    use mp4_atom::Iloc;
-    if meta.get::<Iloc>().is_none() {
-        meta.push(Iloc {
-            item_locations: Vec::new(),
-        });
-    }
-    meta.get_mut::<Iloc>().unwrap().item_locations.push(loc);
 }
 
 /// Build a standalone XMP packet from a bundle of fields. Thin convenience
@@ -1063,20 +670,6 @@ mod tests {
     }
 
     // ── HEIC tests ──────────────────────────────────────────────────────
-
-    /// Check the file extension dispatcher — HEIC paths route to libheif,
-    /// everything else to XmpFile.
-    #[test]
-    fn is_heif_path_recognises_heic_variants() {
-        assert!(is_heif_path(Path::new("/a/b.heic")));
-        assert!(is_heif_path(Path::new("/a/b.HEIC")));
-        assert!(is_heif_path(Path::new("/a/b.HEIF")));
-        assert!(is_heif_path(Path::new("/a/b.hif")));
-        assert!(is_heif_path(Path::new("/a/b.avif")));
-        assert!(!is_heif_path(Path::new("/a/b.jpg")));
-        assert!(!is_heif_path(Path::new("/a/b.mov")));
-        assert!(!is_heif_path(Path::new("/a/b")));
-    }
 
     /// `build_xmp_packet` emits a packet bytes blob that libheif can accept.
     /// Verifies the packet contains the rdf:RDF wrapper and our data.
